requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
yfinance==0.2.18
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Werkzeug==2.3.7
alpha-vantage==2.3.1
pytrends==4.9.2
tweepy==4.14.0
textblob==0.18.0
redis==5.0.8
gunicorn==22.0.0
ratelimit==2.2.1
ortools==9.10.4067
python-dotenv==1.0.0
# MCP and Enhanced Alpaca Trading Dependencies
alpaca-py>=0.39.4
mcp>=1.6.0
fastmcp>=0.1.0